audio/pyproject.toml
[project]
name = "audio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradium>=0.5.7",
    "openai>=1.0.0",
    "python-dotenv>=1.0.1",
    "sounddevice>=0.5.5",
    "websockets>=16.0",
]
